chore: extract `LiveReloadMiddleware` to a separate file
The module still reexports the type at the same place, but the
middleware has been extracted to make quite a big file a little bit
smaller.

Also includes some basic tests for the middleware and tiny doc fixes.
